+++ common/src/main/java/vg/skye/hexstuff/ItemStackableIota.java
@@ -1,0 +1,8 @@
+package vg.skye.hexstuff;
+
+import at.petrak.hexcasting.api.spell.mishaps.Mishap;
+import net.minecraft.world.item.ItemStack;
+
+public interface ItemStackableIota {
+    ItemStack hexstuff$getItemStack() throws Mishap;
+}