+++ common/src/main/java/vg/skye/hexstuff/casting/patterns/display/OpSetTooltip.kt
@@ -9,8 +9,7 @@ import at.petrak.hexcasting.api.spell.mishaps.MishapInvalidIota
 import net.minecraft.network.chat.HoverEvent
 import net.minecraft.world.entity.decoration.ItemFrame
 import net.minecraft.world.entity.item.ItemEntity
-import net.minecraft.world.item.ItemStack
-import ram.talia.hexal.api.spell.iota.MoteIota
+import vg.skye.hexstuff.ItemStackableIota
 import vg.skye.hexstuff.casting.getDisplay
 import vg.skye.hexstuff.casting.iota.DisplayIota
 
@@ -21,13 +20,8 @@ class OpSetTooltip : ConstMediaAction {
         val main = args.getDisplay(0, argc)
         val hover = when (val arg = args[1]) {
             is DisplayIota -> HoverEvent(HoverEvent.Action.SHOW_TEXT, arg.component)
-            is MoteIota -> {
-                val mote = arg.selfOrNull()
-                if (mote == null)
-                    throw MishapInvalidIota.of(arg, 0, "itementityitemframeitemdisplay")
-                val stack = ItemStack(mote.item, mote.count.toInt())
-                stack.tag = mote.tag
-                HoverEvent(HoverEvent.Action.SHOW_ITEM, HoverEvent.ItemStackInfo(stack))
+            is ItemStackableIota -> {
+                HoverEvent(HoverEvent.Action.SHOW_ITEM, HoverEvent.ItemStackInfo(arg.`hexstuff$getItemStack`()))
             }
             is EntityIota -> {
                 if (!ctx.isEntityInRange(arg.entity))