+++ forge/src/main/java/vg/skye/hexstuff/forge/HexStuffAbstractionsImpl.java
@@ -1,5 +1,6 @@
 package vg.skye.hexstuff.forge;
 
+import net.minecraftforge.fml.ModList;
 import vg.skye.hexstuff.HexStuffAbstractions;
 import net.minecraftforge.fml.loading.FMLPaths;
 
@@ -15,5 +16,9 @@ public class HexStuffAbstractionsImpl {
 	
     public static void initPlatformSpecific() {
         HexStuffConfigForge.init();
+    }
+
+    public static boolean isModLoaded(String id) {
+        return ModList.get().isLoaded(id);
     }
 }